refactor(core): extract SpeechPipeline into core.speech module
The persistent text-to-speech pipeline had accreted inside the EasySpeak
god-object, mixing subprocess lifecycle management in with the wake-word
loop and plugin plumbing. We pull it into a dedicated core.speech module
so the speech concern owns its own state and can be reasoned about and
tested in isolation.

EasySpeak composes the pipeline rather than inheriting its mechanics, and
keeps speak() as a thin delegate so the plugin-facing API is untouched.
The pipeline's tests move alongside it. No behaviour change.

Also lands the previously-deferred PortAudio noise suppression here, where
it belongs with the rest of the audio-subsystem plumbing rather than as a
standalone patch.
